feat: added fetch collections only after the complete your guide
