+++ apps/@sparrow-desktop/src/pages/workspace-page/sub-pages/RestExplorer/RestExplorerPage.ViewModel.ts
@@ -3884,108 +3884,6 @@ class RestExplorerViewModel {
     return;
   };
 
-  /**
-   * Fetch collections from services and insert to repository
-   * @param workspaceId - id of current workspace
-   */
-  public fetchCollections = async (
-    workspaceId: string,
-  ): Promise<{ collectionItemTabsToBeDeleted?: string[] }> => {
-    const isGuestUser = await this.getGuestUserState();
-    if (!workspaceId || isGuestUser) {
-      return {};
-    }
-
-    const getCollectionItemIds = (
-      collectionItem: any,
-      collectedIds: string[],
-    ): void => {
-      const stack = [collectionItem];
-      while (stack.length > 0) {
-        const item = stack.pop();
-        if (!item) continue;
-
-        if (!item.type) {
-          // Collection
-          collectedIds.push(item._id);
-        } else {
-          // Folder, Http Request, WebSocket Request
-          collectedIds.push(item.id);
-        }
-
-        if (Array.isArray(item.items)) {
-          stack.push(...item.items);
-        }
-      }
-    };
-
-    const baseUrl = await this.constructBaseUrl(workspaceId);
-    const workspaceData =
-      await this.workspaceRepository.readWorkspace(workspaceId);
-
-    let res;
-    if (
-      workspaceData &&
-      workspaceData.workspaceType === WorkspaceType.PUBLIC &&
-      workspaceData.isShared
-    ) {
-      res = await this.collectionService.fetchPublicCollection(
-        workspaceId,
-        constants.API_URL,
-      );
-    } else {
-      res = await this.collectionService.fetchCollection(workspaceId, baseUrl);
-    }
-
-    if (!res?.isSuccessful || !res?.data?.data) {
-      return {};
-    }
-
-    const collections = res.data.data;
-    const processedCollections: any[] = [];
-    const collectionIds: string[] = [];
-
-    const chunkSize = 100;
-    for (let i = 0; i < collections.length; i += chunkSize) {
-      const chunk = collections.slice(i, i + chunkSize);
-      for (const col of chunk) {
-        const collection = createDeepCopy(col);
-        collection.workspaceId = workspaceId;
-        collection.id = col._id;
-        if (!collection.description) collection.description = "";
-        delete collection._id;
-
-        processedCollections.push(collection);
-        collectionIds.push(col._id);
-      }
-      await new Promise((res) => setTimeout(res));
-    }
-
-    await this.collectionRepository.bulkInsertData(
-      workspaceId,
-      processedCollections,
-    );
-    await this.collectionRepository.deleteOrphanCollections(
-      workspaceId,
-      collectionIds,
-    );
-
-    const collectionItemIds: string[] = [];
-    for (const collection of collections) {
-      getCollectionItemIds(collection, collectionItemIds);
-    }
-
-    const collectionItemTabsToBeDeleted =
-      await this.tabRepository.getIdOfTabsThatDoesntExistAtCollectionLevel(
-        workspaceId,
-        collectionItemIds,
-      );
-
-    return {
-      collectionItemTabsToBeDeleted,
-    };
-  };
-
   public InsertGenerateTrialFlow = async (collectionId: string) => {
     try {
       const response =
